requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
Pillow>=10.0.0
gTTS>=2.5.0
python-pptx>=0.6.23
